Adds new arguments to rotate sample and compute only cubemap
--cubemap-only
It only compute cubemap ggx without the panorama

--sample-rotation {integer}
Rotate the samples so if you specify 16 the number of rotation will be 16
with an angle of 2*PI/16. There is a kind of random offset used to avoid
sampling pattern. The number of rotations is used to compute background and ggx

In term of quality with 2048 samples and 18 rotations you have better quality
than 65k samples without rotations. For a good quality:
ggx -> 1024 samples and 18 rotations
background -> 128 samples and 18 rotations